+++ gore-and-ash-sync.cabal
@@ -1,5 +1,5 @@
 name:                gore-and-ash-sync
-version:             1.2.0.1
+version:             2.0.0.0
 synopsis:            Gore&Ash module for high level network synchronization
 description:         Please see README.md
 homepage:            https://github.com/Teaspot-Studio/gore-and-ash-sync
@@ -16,50 +16,38 @@ extra-source-files:
   CHANGELOG.md
   stack.yaml
 
+flag examples
+  description: Enable building of examples
+  default:     False
+
 library
   hs-source-dirs:      src
   exposed-modules:
-                       Game.GoreAndAsh.Sync
-                       Game.GoreAndAsh.Sync.API
-                       Game.GoreAndAsh.Sync.Message
-                       Game.GoreAndAsh.Sync.Module
-                       Game.GoreAndAsh.Sync.Remote
-                       Game.GoreAndAsh.Sync.Remote.Actor
-                       Game.GoreAndAsh.Sync.Remote.Collection
-                       Game.GoreAndAsh.Sync.Remote.Sync
-                       Game.GoreAndAsh.Sync.State
+    Game.GoreAndAsh.Sync
+    Game.GoreAndAsh.Sync.API
+    Game.GoreAndAsh.Sync.Module
+    Game.GoreAndAsh.Sync.State
 
   default-language:    Haskell2010
-  build-depends:       base >= 4.7 && < 5
-                     , bytestring >= 0.10.6.0
-                     , cereal >= 0.4.1.1
-                     , containers >= 0.5.6.2
-                     , deepseq >= 1.4.1.1
-                     , exceptions >= 0.8.0.2
-                     , gore-and-ash >= 1.1.0.0
-                     , gore-and-ash-actor >= 1.1.0.0
-                     , gore-and-ash-logging >= 2.0.0.0
-                     , gore-and-ash-network >= 1.4.0.0
-                     , hashable >= 1.2.3.3
-                     , mtl >= 2.2.1
-                     , text >= 1.2.2.0
-                     , unordered-containers >= 0.2.5.1
+  build-depends:
+     base                     >= 4.7      && < 5
+   , gore-and-ash             >= 2.0      && < 2.1
+   , gore-and-ash-logging     >= 3.0      && < 3.1
+   , gore-and-ash-network     >= 2.0      && < 2.1
 
   default-extensions:
-                       Arrows
-                       BangPatterns
-                       ConstraintKinds
-                       DeriveGeneric
-                       FlexibleContexts
-                       FlexibleInstances
-                       FunctionalDependencies
-                       GADTs
-                       GeneralizedNewtypeDeriving
-                       MultiParamTypeClasses
-                       OverloadedStrings
-                       RecordWildCards
-                       ScopedTypeVariables
-                       StandaloneDeriving
-                       TupleSections
-                       TypeFamilies
-                       UndecidableInstances
+    BangPatterns
+    DeriveGeneric
+    FlexibleContexts
+    FlexibleInstances
+    FunctionalDependencies
+    GADTs
+    GeneralizedNewtypeDeriving
+    MultiParamTypeClasses
+    OverloadedStrings
+    RecordWildCards
+    ScopedTypeVariables
+    StandaloneDeriving
+    TupleSections
+    TypeFamilies
+    UndecidableInstances
