+++ src_old/Game/GoreAndAsh/Sync/Message.hs
