+++ stack.yaml
@@ -1,7 +1,7 @@
 # For more information, see: https://github.com/commercialhaskell/stack/blob/release/doc/yaml_configuration.md
 
 # Specifies the GHC version and set of packages available (e.g., lts-3.5, nightly-2015-09-21, ghc-7.10.2)
-resolver: lts-7.9
+resolver: lts-7.12
 
 # Local packages, usually specified by relative directory name
 packages:
@@ -11,12 +11,11 @@ packages:
 #    commit: d71d9d24f20c5d819e9397cbb0098cbc6ef01f88
 
 # Packages to be pulled from upstream that are not in the resolver (e.g., acme-missiles-0.3)
-extra-deps: 
+extra-deps:
 - typesafe-endian-0.1.0.1
-- gore-and-ash-1.2.2.0
-- gore-and-ash-actor-1.2.2.0
-- gore-and-ash-logging-2.0.1.0
-- gore-and-ash-network-1.4.0.0
+- gore-and-ash-2.0.0.0
+- gore-and-ash-logging-3.0.0.0
+- gore-and-ash-network-2.0.0.0
 
 # Override default flag values for local packages and extra-deps
 flags: {}
